Rename ideal directory
