+++ s2e-ff/src/components/ideal/InitializeRelativeAttitudeController.hpp
