+++ s2e-ff/src/components/ideal/RelativeAttitudeController.hpp
