+++ s2e-ff/src/simulation/spacecraft/ff_components.hpp
@@ -8,10 +8,10 @@
 #include <components/ideal/force_generator.hpp>
 #include <components/real/cdh/on_board_computer.hpp>
 
-#include "../../components/IdealComponents/InitializeRelativeAttitudeController.hpp"
 #include "../../components/aocs/InitializeRelativeDistanceSensor.hpp"
 #include "../../components/aocs/InitializeRelativePositionSensor.hpp"
 #include "../../components/aocs/InitializeRelativeVelocitySensor.hpp"
+#include "../../components/ideal/InitializeRelativeAttitudeController.hpp"
 
 class FfComponents : public InstalledComponents {
  public:
