+++ s2e-ff/src/simulation/spacecraft/ff_components_2.hpp
@@ -8,9 +8,9 @@
 #include <components/ideal/force_generator.hpp>
 #include <components/real/cdh/on_board_computer.hpp>
 
-#include "../../components/IdealComponents/InitializeRelativeAttitudeController.hpp"
 #include "../../components/aocs/InitializeRelativeDistanceSensor.hpp"
 #include "../../components/aocs/InitializeRelativePositionSensor.hpp"
+#include "../../components/ideal/InitializeRelativeAttitudeController.hpp"
 
 class FfComponents2 : public InstalledComponents {
  public:
